Refactor VirtualMachine.cs
